requirements.txt
streamlit
pandas
plotly
gspread
google-auth
Pillow
